Move math operations from wrapper to ops
