+++ luchador/nn/core/backend/tensorflow/wrapper.py
@@ -64,116 +64,6 @@ class TensorMixin(object):  # pylint: disable=too-few-public-methods
         _other = self._extract_operand(other)
         return Tensor(tensor=_other//self._tensor, name=name)
 
-    def mean(self, axis=None, keep_dims=False, name=None):
-        """Compute mean across the given axis
-
-        Parameters
-        ----------
-        axis : int, list or None
-            The dimensions to compute mean. If None (the default),
-            reduces all dimensions.
-        keep_dims: bool
-            If true, retains reduced dimensions with length 1.
-        name: str
-            A name for the operation.
-
-        Returns
-        -------
-        Tensor
-            The resulting Tensor
-        """
-        _tensor = tf.reduce_mean(
-            self.unwrap(), axis=axis, keep_dims=keep_dims, name=name)
-        return Tensor(tensor=_tensor, name=name)
-
-    def sum(self, axis=None, keep_dims=False, name=None):
-        """Compute sum across the given axis
-
-        Parameters
-        ----------
-        axis : int, list or None
-            The dimensions to compute sum. If None (the default),
-            reduces all dimensions.
-        keep_dims: bool
-            If true, retains reduced dimensions with length 1.
-        name: str
-            A name for the operation.
-
-        Returns
-        -------
-        Tensor
-            The resulting Tensor
-        """
-        _tensor = tf.reduce_sum(
-            self.unwrap(), axis=axis, keep_dims=keep_dims, name=name)
-        return Tensor(tensor=_tensor, name=name)
-
-    def max(self, axis=None, keep_dims=False, name=None):
-        """Compute max across the given axis
-
-        Parameters
-        ----------
-        axis : int, list or None
-            The dimensions to compute max. If None (the default),
-            reduces all dimensions.
-        keep_dims: bool
-            If true, retains reduced dimensions with length 1.
-        name: str
-            A name for the operation.
-
-        Returns
-        -------
-        Tensor
-            The resulting Tensor
-        """
-        _tensor = tf.reduce_max(
-            self.unwrap(), axis=axis, keep_dims=keep_dims, name=name)
-        return Tensor(tensor=_tensor, name=name)
-
-    def reshape(self, new_shape, name=None):
-        """Reshape tensor.
-
-        Parameters
-        ----------
-        new_shape : tuple
-            new shape
-
-        name : str
-            Name of operation
-
-        Returns
-        -------
-        Tensor
-            Tensor with new shape
-        """
-        _tensor = tf.reshape(self.unwrap(), shape=new_shape)
-        return Tensor(tensor=_tensor, name=name)
-
-    def tile(self, pattern, name=None):
-        """Tile tensor.
-
-        Parameters
-        ----------
-        pattern : tuple
-            tile pattern
-
-        Notes
-        -----
-        Currently only constant pattern is allowed.
-        """
-        if not luchador.util.is_iteratable(pattern):
-            raise ValueError('`pattern` must be iteratable')
-        pattern = tuple(pattern)
-
-        if len(pattern) > self.n_dim:
-            prepend = (1, ) * (len(pattern) - self.n_dim)
-            tensor = self.reshape(prepend + self.shape).unwrap()
-        else:
-            prepend = (1, ) * (self.n_dim - len(pattern))
-            pattern = prepend + pattern
-            tensor = self.unwrap()
-        return Tensor(tf.tile(tensor, pattern, name), name=name)
-
 
 def _get_dtype_str(tensor):
     return tensor.dtype.as_numpy_dtype().dtype.name
