+++ tests/unit/nn/layer/anonymous_test.py
@@ -95,43 +95,43 @@ class AnonymousSingleInputTest(TestCase):
         """Anonymous layer can handle mean"""
         input_val = np.random.rand(3, 4, 5, 6)
         output_val = input_val.mean(axis=1)
-        _test('x.mean(axis=1)', input_val, output_val, self.get_scope())
+        _test('mean(x, axis=1)', input_val, output_val, self.get_scope())
 
     def test_mean_multi(self):
         """Anonymous layer can handle cos"""
         input_val = np.random.rand(3, 4, 5, 6)
         output_val = np.mean(input_val, axis=(1, 2))
-        _test('x.mean(axis=(1, 2))', input_val, output_val, self.get_scope())
+        _test('mean(x, axis=(1, 2))', input_val, output_val, self.get_scope())
 
     def test_mean_all(self):
         """Anonymous layer can handle cos"""
         input_val = np.random.rand(3, 4, 5, 6)
         output_val = np.mean(input_val)
-        _test('x.mean()', input_val, output_val, self.get_scope())
+        _test('mean(x)', input_val, output_val, self.get_scope())
 
     def test_reshape(self):
         """Anonymous layer can handle reshape"""
         input_val = np.random.rand(3, 4)
         output_val = input_val.reshape((-1, 1))
-        _test('x.reshape((-1, 1))', input_val, output_val, self.get_scope())
+        _test('reshape(x, (-1, 1))', input_val, output_val, self.get_scope())
 
     def test_tile(self):
         """Anonymous layer can handle tile"""
         input_val = np.random.rand(3, 4)
         output_val = np.tile(input_val, (1, 3, 5))
-        _test('x.tile((1, 3, 5))', input_val, output_val, self.get_scope())
+        _test('tile(x, (1, 3, 5))', input_val, output_val, self.get_scope())
 
     def test_max(self):
         """Anonymous layer can handle max"""
         input_val = np.random.rand(3, 4)
         output_val = input_val.max(axis=1)
-        _test('x.max(axis=1)', input_val, output_val, self.get_scope())
+        _test('max(x, axis=1)', input_val, output_val, self.get_scope())
 
     def test_sum(self):
         """Anonymous layer can handle sum"""
         input_val = np.random.rand(3, 4)
         output_val = input_val.sum(axis=1)
-        _test('x.sum(axis=1)', input_val, output_val, self.get_scope())
+        _test('sum(x, axis=1)', input_val, output_val, self.get_scope())
 
     def test_mean_shift(self):
         """Anonymous layer can handle complex arithmetic"""
@@ -139,7 +139,7 @@ class AnonymousSingleInputTest(TestCase):
         mean_ = np.tile(input_val.mean(axis=1, keepdims=True), (1, 4))
         output_val = input_val - mean_
         _test(
-            'x - x.mean(axis=1, keep_dims=True).tile((1, 4))',
+            'x - tile(mean(x, axis=1, keep_dims=True), (1, 4))',
             input_val, output_val, self.get_scope()
         )
 
@@ -147,4 +147,4 @@ class AnonymousSingleInputTest(TestCase):
         """Anonymous layer can handle complex arithmetic"""
         input_val = np.random.rand(3, 4)
         output_val = (input_val * input_val).sum()
-        _test('(x * x).sum()', input_val, output_val, self.get_scope())
+        _test('sum(x * x)', input_val, output_val, self.get_scope())
