+++ tests/unit/nn/optimizer_test.py
@@ -13,7 +13,6 @@ from luchador import nn
 from tests.unit import fixture
 
 BE = luchador.get_nn_backend()
-
 # pylint: disable=too-many-locals, invalid-name
 
 
@@ -101,7 +100,7 @@ class OptimizerGradientTest(fixture.TestCase):
             y = x * x / 2
             grads_and_vars = [
                 (nn.ops.clip_by_value(grad, max_value=1, min_value=-1), var)
-                for grad, var in sgd.compute_gradients(y.sum(), wrt=x)
+                for grad, var in sgd.compute_gradients(nn.ops.sum(y), wrt=x)
             ]
             op = sgd.apply_gradients(grads_and_vars)
